parametrise the package
