+++ src/main/g8/default.properties
@@ -1,4 +1,5 @@
 name=my-awesome-lambda
 stack=mobile
 deployBucket=mobile-dist
+package=com.gu.$name$
 aws_lambda_java_core_version=maven(com.amazonaws, aws-lambda-java-core)
